Fixed double python issue
git-svn-id: http://compss.bsc.es/svn/compss/framework/trunk@2331 9ab3861e-6c05-4e1b-b5ef-99af60850597

+++ compss/programming_model/bindings/python/src/pycompss/runtime/binding.py
@@ -30,7 +30,7 @@ from cPickle import PicklingError
 
 python_to_compss = {types.IntType: Type.INT,          # int
                     types.LongType: Type.LONG,        # long
-                    types.FloatType: Type.FLOAT,      # float
+                    types.FloatType: Type.DOUBLE,      # float
                     types.BooleanType: Type.BOOLEAN,  # bool
                     types.StringType: Type.STRING,    # str
                     # The type of instances of user-defined classes
